Fixed instr regression

+++ src/vtlengine/duckdb_transpiler/sql/init.sql
@@ -271,7 +271,7 @@ CREATE OR REPLACE MACRO vtl_period_to_natural(input VARCHAR) AS (
 
 -- VTL instr(string, pattern, start, occurrence)
 CREATE OR REPLACE MACRO vtl_instr(
-    s VARCHAR, pat VARCHAR, start_pos_raw INTEGER, occur_raw INTEGER
+    s VARCHAR, pat VARCHAR, start_pos_raw BIGINT, occur_raw BIGINT
 ) AS (
     CASE
         WHEN s IS NULL THEN NULL
